+++ src/elements/index.js
@@ -21,6 +21,7 @@ import ItineraryDisplay from './ItineraryDisplay';
 import ExtendedBackground from './ExtendedBackground';
 import Fieldset from './Fieldset';
 import Review from './Review'
+import Stars from './Stars'
 
 export {
   Button,
@@ -45,5 +46,6 @@ export {
   ItineraryDisplay,
   ExtendedBackground,
   Fieldset,
-  Review
+  Review,
+  Stars
 };
